+++ library/sbi-rt/CHANGELOG.md
@@ -17,6 +17,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - pmu: change `counter_idx_mask` and `counter_idx_base` parameters into `counter_idx` with type `CounterMask`.
 - Migrate sbi-rt crate to Rust 2024 edition.
+- doc: update crate-level documentation to specify that all SBI call functions return `SbiRet<usize>`.
 
 ### Fixed